Add back Char CNN for QANet
